requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
plotly>=5.15.0
seaborn>=0.12.0
folium>=0.14.0
streamlit-folium>=0.13.0
argopy>=0.1.13
warnings
datetime
pinecone-client>=3.0.0
sentence-transformers>=2.2.2
groq>=0.4.1